requirements.txt
fastapi
uvicorn
firebase-admin
apscheduler
requests
streamlit
gTTS
googletrans==4.0.0-rc1
python-dotenv